requirements.txt
python-decouple~=3.8
fastapi~=0.96.0
uvicorn~=0.22.0
SQLAlchemy~=2.0.15
psycopg2-binary~=2.9.6
pydub~=0.25.1
python-multipart~=0.0.6
requests